+++ pages/Content-Marketing.js
@@ -1,9 +1,9 @@
 import Layout from "../components/Layout";
 
-export default function Blog() {
+export default function ContentMarketing() {
   return (
-    <Layout title="Blog">
-      <h1>Blog</h1>
+    <Layout title="Content Marketing">
+      <h1>Content Marketing</h1>
     </Layout>
   );
 }
